requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
pandas==2.3.2
numpy==2.0.2
faiss-cpu==1.12.0
sentence-transformers==5.1.0
google-generativeai==0.8.5